+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# [1.38.0](https://github.com/deriv-com/deriv-analytics/compare/v1.37.0...v1.38.0) (2026-02-11)
+
+### Bug Fixes
+
+- release issue ([361b92f](https://github.com/deriv-com/deriv-analytics/commit/361b92f68fca7bb8b95465af0e7ee808a0d0ef5b))
+- release issue ([d32cfd6](https://github.com/deriv-com/deriv-analytics/commit/d32cfd6611385158112341428500167a7326733a))
+- release issue ([9638994](https://github.com/deriv-com/deriv-analytics/commit/96389941aa4dc9e56df02aa25e4cc67bea004738))
+
+### Features
+
+- add configurable PostHog api_host and enhance security ([8d58821](https://github.com/deriv-com/deriv-analytics/commit/8d58821a6adb84c54f04a4fb9b2c202d9f3e1deb))
+
+## [1.38.1](https://github.com/deriv-com/deriv-analytics/compare/v1.38.0...v1.38.1) (2026-02-11)
+
+### Bug Fixes
+
+- release issue ([d32cfd6](https://github.com/deriv-com/deriv-analytics/commit/d32cfd6611385158112341428500167a7326733a))
+- release issue ([9638994](https://github.com/deriv-com/deriv-analytics/commit/96389941aa4dc9e56df02aa25e4cc67bea004738))
+
+# [1.38.0](https://github.com/deriv-com/deriv-analytics/compare/v1.37.0...v1.38.0) (2026-02-11)
+
+### Features
+
+- add configurable PostHog api_host and enhance security ([8d58821](https://github.com/deriv-com/deriv-analytics/commit/8d58821a6adb84c54f04a4fb9b2c202d9f3e1deb))
+
 # [1.37.0](https://github.com/deriv-com/deriv-analytics/compare/v1.36.0...v1.37.0) (2026-02-11)
 
 ### Bug Fixes